# smart/openapi.yaml
openapi: 3.0.3
info:
  title: PrayCalc Smart API
  version: 1.0.0
  description: |
    REST API for PrayCalc prayer time calculations, webhook notifications,
    smart home integrations (Google Home, Alexa, Home Assistant), and billing.

    **Base URL:** `https://api.praycalc.com`

    **Authentication:** Prayer times are public. Webhook management, billing,
    and smart home account linking require a Bearer JWT from Hasura Auth
    (obtained via https://auth.ummat.dev).

    **Rate Limits:** 100 requests/min for authenticated users, 30 requests/min
    for anonymous users.
  contact:
    name: PrayCalc
    url: https://praycalc.com
    email: support@praycalc.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.praycalc.com
    description: Production
  - url: http://localhost:4010
    description: Local development
paths:
...
  /billing/checkout:
    post:
      operationId: createCheckout
      summary: Create Stripe checkout session
      description: |
        Creates a Stripe Checkout session for an Ummat+ subscription.
        Returns a URL to redirect the user to Stripe's hosted checkout page.
      tags:
        - Billing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: Customer email for Stripe
                  example: user@example.com
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                    description: Stripe-hosted checkout URL
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Stripe API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /billing/status:
    get:
      operationId: getSubscriptionStatus
      summary: Check subscription status
      description: Returns the current subscription plan and status for the authenticated user.
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: string
                    enum:
                      - free
                      - plus
                    example: free
                  status:
                    type: string
                    enum:
                      - none
                      - active
                      - canceled
                      - past_due
                    example: none
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /billing/portal:
    post:
      operationId: createPortal
      summary: Create Stripe customer portal session
      description: |
        Creates a Stripe Customer Portal session where users can manage
        their subscription (cancel, update payment method, view invoices).
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Portal session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No subscription found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /billing/verify-receipt:
    post:
      operationId: verifyReceipt
      summary: Verify mobile IAP receipt
      description: |
        Verifies an iOS or Android in-app purchase receipt and activates
        the subscription. Used by the Flutter mobile app.
      tags:
        - Billing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - platform
                - receipt
                - productId
              properties:
                platform:
                  type: string
                  enum:
                    - ios
                    - android
                receipt:
                  type: string
                  description: Platform-specific receipt data
                productId:
                  type: string
                  description: Product identifier (e.g., ummat_plus_yearly)
                  example: ummat_plus_yearly
      responses:
        "200":
          description: Receipt verified, subscription activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: active
                  plan:
                    type: string
                    example: plus
        "400":
          description: Missing fields or unknown product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Hasura Auth JWT containing `x-hasura-user-id` and `x-hasura-default-role`
        claims. Obtain from `POST https://auth.ummat.dev/signin/email-password`.
  schemas:
...
    Error:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error code or short message
          example: "Invalid coordinates"
        message:
          type: string
          description: Detailed error description
          example: "lat must be -90..90, lng must be -180..180"
...